+++ src/main/java/com/example/project/user/controller/UserInfoController.java
@@ -8,10 +8,14 @@ import com.example.project.user.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
+import java.util.concurrent.ExecutionException;
 
+@Slf4j
 @RestController
 @RequestMapping("/api/user/info")
 @RequiredArgsConstructor
@@ -33,5 +37,22 @@ public class UserInfoController {
             UserResponse updateResponse = userService.updateUser(new UserDetail(userResponse.toEntity()), updateRequest);
         }
         response.sendRedirect("/localhost:8080/user/info");
+    }
+
+    //TODO 토큰 인증으로 마이페이지 접속할 수 있도록 추가하기
+    @GetMapping("/my-page/{id}")
+    public void myPage(Model model, @PathVariable Long id) {
+        var userResponse = userService.find(id);
+        log.info("[ SYSTEM ] MyPage user 조회 성공했습니다 {}", id);
+        model.addAttribute("UserInfo", userResponse);
+    }
+
+    //TODO 토큰 인증으로 유저 정보 수정 가능하도록 만들기
+    @GetMapping("/edit/{id}")
+    public void editInfo(Model model, @PathVariable Long id) {
+        var userResponse = userService.find(id);
+        log.info("[ SYSTEM ] Edit user 조회 성공했습니다 {}", id);
+        model.addAttribute("UserResponse", userResponse);
+        model.addAttribute("UpdateRequest", new UserUpdateRequest());
     }
 }
