+++ src/main/java/com/example/project/user/controller/UserLoginController.java
@@ -10,6 +10,7 @@ import com.example.project.user.service.CookieService;
 import com.example.project.user.service.SessionService;
 import com.example.project.user.service.LoginService;
 import com.example.project.user.service.UserService;
+import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -47,7 +48,6 @@ public class UserLoginController {
             if (cookieValue == null) userResponse = noCookieLogin(loginRequest);
             else userResponse = cookieLogin(loginRequest, cookieValue, request);
 
-
             sessionService.sessionRegistration(request, userResponse);
             response.addCookie(cookieService.authCookieIssue(loginRequest));
             //TODO loginInfo cookie 반환하도록 수정
@@ -80,39 +80,5 @@ public class UserLoginController {
         }
 
         return userResponse;
-    }
-
-
-    //TODO 토큰 인증으로 마이페이지 접속할 수 있도록 추가하기
-    @GetMapping("/my-page/{id}")
-    public String myPage(Model model, @PathVariable Long id) throws ExecutionException, InterruptedException {
-        var userResponse = userService.find(id);
-        log.info("[ SYSTEM ] MyPage user 조회 성공했습니다 {}", id);
-        model.addAttribute("UserInfo", userResponse);
-
-        return "/authentication/user/info/info";
-    }
-
-    //TODO 토큰 인증으로 유저 정보 수정 가능하도록 만들기
-    @GetMapping("/edit/{id}")
-    public String editInfo(Model model, @PathVariable Long id) throws ExecutionException, InterruptedException {
-        var userResponse = userService.find(id);
-        log.info("[ SYSTEM ] Edit user 조회 성공했습니다 {}", id);
-        model.addAttribute("UserResponse", userResponse);
-        model.addAttribute("UpdateRequest", new UserUpdateRequest());
-
-        return "/authentication/user/info/edit_info";
-    }
-
-    //TODO 토큰 인증으로 유저 정보 수정하도록 만들기
-    @PostMapping("/edit/{id}")
-    public String edit(UserUpdateRequest request, Model model, @PathVariable Long id) throws ExecutionException, InterruptedException {
-        var userResponse = userService.find(id);
-        UserDetail userDetail = new UserDetail(userResponse.toEntity());
-        var edit = userService.updateUser(userDetail, request);
-
-        model.addAttribute("UserInfo", edit);
-
-        return "/authentication/user/info/info";
     }
 }
